API's Folder Structure implementation.
