+++ app/controllers/api_controller.rb
@@ -1,0 +1,6 @@
+class ApiController < ActionController::API
+  # include TokenUserPermission
+  include ApiExceptionHandler
+  include ApiResponse
+  # before_action :authorize_user!
+end
