+++ app/controllers/concerns/api_exception_handler.rb
@@ -1,0 +1,13 @@
+module ApiExceptionHandler
+  extend ActiveSupport::Concern
+
+  included do
+    rescue_from ActiveRecord::RecordNotFound do |e|
+      json_response(message: e.message, status: :not_found)
+    end
+
+    def authenticate_user!(options = {})
+      json_response(message: I18n.t('devise.failure.unauthenticated'), status: :unauthorized) unless signed_in?
+    end
+  end
+end
